genealogy/backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Database & ORM
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# Multi-tenancy
django-tenants>=4.0

# Redis & Caching
django-redis>=5.4.0
redis>=5.0.0

# Async Tasks
celery>=5.3.0
django-celery-beat>=2.5.0

# API Documentation
drf-spectacular>=0.27.0

# Security
django-ratelimit>=4.1.0
django-axes>=6.0.0
bleach>=6.1.0

# Utilities
Pillow>=10.0.0
python-dateutil>=2.8.2
uuid>=1.30

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Logging & Monitoring
sentry-sdk>=1.40.0
django-log-request-id>=2.1.0

# AI Integration
openai>=1.0.0
anthropic>=0.20.0
google-cloud-vision>=1.0.0
Pillow>=10.0.0
